Move helper functions to app utils file
